Relocate result stat clamp inline
Keep the inferred shared body in the dedicated gm inline header rather than assigning it to gm_17AD.h.
